INS-3122: move pulse package to top-level
