+++ network/consensus/adapters/roundstrategy.go
@@ -54,9 +54,9 @@ import (
 	"context"
 	"math/rand"
 
-	"github.com/insolar/insolar/network/consensus/common/pulse"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api/profiles"
+	"github.com/insolar/insolar/pulse"
 
 	"github.com/insolar/insolar/instrumentation/inslogger"
 )
