+++ network/consensus/gcpv2/api/census/census.go
@@ -54,9 +54,9 @@ import (
 	"context"
 
 	"github.com/insolar/insolar/network/consensus/common/cryptkit"
-	"github.com/insolar/insolar/network/consensus/common/pulse"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api/profiles"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api/proofs"
+	"github.com/insolar/insolar/pulse"
 )
 
 type Pulse interface {
