+++ network/consensus/gcpv2/api/census/stored_registries.go
@@ -53,10 +53,10 @@ package census
 import (
 	"github.com/insolar/insolar/network/consensus/common/cryptkit"
 	"github.com/insolar/insolar/network/consensus/common/endpoints"
-	"github.com/insolar/insolar/network/consensus/common/pulse"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api/misbehavior"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api/profiles"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api/proofs"
+	"github.com/insolar/insolar/pulse"
 )
 
 //go:generate minimock -i github.com/insolar/insolar/network/consensus/gcpv2/api/census.VersionedRegistries -o . -s _mock.go -g
