+++ network/consensus/gcpv2/api/census/versioned_registries_mock.go
@@ -8,7 +8,8 @@ import (
 	mm_time "time"
 
 	"github.com/gojuno/minimock"
-	"github.com/insolar/insolar/network/consensus/common/pulse"
+
+	"github.com/insolar/insolar/pulse"
 )
 
 // VersionedRegistriesMock implements VersionedRegistries
