+++ network/consensus/gcpv2/api/chronicles.go
@@ -52,9 +52,9 @@ package api
 
 import (
 	"github.com/insolar/insolar/network/consensus/common/cryptkit"
-	"github.com/insolar/insolar/network/consensus/common/pulse"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api/census"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api/profiles"
+	"github.com/insolar/insolar/pulse"
 )
 
 type ConsensusChronicles interface {
