+++ network/consensus/gcpv2/api/profiles/candidate_profile_mock.go
@@ -8,11 +8,12 @@ import (
 	mm_time "time"
 
 	"github.com/gojuno/minimock"
+
 	"github.com/insolar/insolar/insolar"
 	"github.com/insolar/insolar/network/consensus/common/cryptkit"
 	"github.com/insolar/insolar/network/consensus/common/endpoints"
-	"github.com/insolar/insolar/network/consensus/common/pulse"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api/member"
+	"github.com/insolar/insolar/pulse"
 )
 
 // CandidateProfileMock implements CandidateProfile
