+++ network/consensus/gcpv2/api/proofs/payloads.go
@@ -52,7 +52,7 @@ package proofs
 
 import (
 	"github.com/insolar/insolar/network/consensus/common/longbits"
-	"github.com/insolar/insolar/network/consensus/common/pulse"
+	"github.com/insolar/insolar/pulse"
 )
 
 type OriginalPulsarPacket interface {
