+++ network/consensus/gcpv2/api/transport/packet_parser_mock.go
@@ -7,10 +7,11 @@ import (
 	mm_time "time"
 
 	"github.com/gojuno/minimock"
+
 	"github.com/insolar/insolar/insolar"
 	"github.com/insolar/insolar/network/consensus/common/cryptkit"
-	"github.com/insolar/insolar/network/consensus/common/pulse"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api/phases"
+	"github.com/insolar/insolar/pulse"
 )
 
 // PacketParserMock implements PacketParser
