+++ network/consensus/gcpv2/censusimpl/census_builder.go
@@ -57,10 +57,10 @@ import (
 	"github.com/insolar/insolar/instrumentation/inslogger"
 
 	"github.com/insolar/insolar/insolar"
-	"github.com/insolar/insolar/network/consensus/common/pulse"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api/census"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api/profiles"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api/proofs"
+	"github.com/insolar/insolar/pulse"
 )
 
 func newLocalCensusBuilder(ctx context.Context, chronicles *localChronicles, pn pulse.Number,
