+++ network/consensus/gcpv2/censusimpl/census_templates_test.go
@@ -60,7 +60,7 @@ import (
 
 	"github.com/insolar/insolar/network/consensus/common/cryptkit"
 
-	"github.com/insolar/insolar/network/consensus/common/pulse"
+	"github.com/insolar/insolar/pulse"
 
 	"github.com/insolar/insolar/insolar"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api/profiles"
