+++ network/consensus/gcpv2/censusimpl/local_chronicle.go
@@ -57,9 +57,9 @@ import (
 	"github.com/insolar/insolar/network/consensus/common/cryptkit"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api/profiles"
 
-	"github.com/insolar/insolar/network/consensus/common/pulse"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api/census"
+	"github.com/insolar/insolar/pulse"
 )
 
 func NewLocalChronicles(profileFactory profiles.Factory) LocalConsensusChronicles {
