+++ network/consensus/gcpv2/censusimpl/local_chronicle_test.go
@@ -53,7 +53,7 @@ package censusimpl
 import (
 	"testing"
 
-	"github.com/insolar/insolar/network/consensus/common/pulse"
+	"github.com/insolar/insolar/pulse"
 
 	"github.com/insolar/insolar/network/consensus/gcpv2/api/census"
