+++ network/consensus/gcpv2/core/round_worker.go
@@ -52,14 +52,15 @@ package core
 
 import (
 	"context"
-	"github.com/insolar/insolar/network/consensus/common/capacity"
 	"sync/atomic"
 	"time"
 
+	"github.com/insolar/insolar/network/consensus/common/capacity"
+
 	"github.com/insolar/insolar/instrumentation/inslogger"
-	"github.com/insolar/insolar/network/consensus/common/pulse"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api/census"
+	"github.com/insolar/insolar/pulse"
 )
 
 const (
