+++ network/consensus/serialization/globula_protocol_test.go
@@ -58,11 +58,11 @@ import (
 
 	"github.com/insolar/insolar/network/consensus/adapters"
 	"github.com/insolar/insolar/network/consensus/common/longbits"
-	"github.com/insolar/insolar/network/consensus/common/pulse"
 	"github.com/insolar/insolar/network/consensus/gcpv2/api/phases"
 	"github.com/insolar/insolar/network/hostnetwork/host"
 	"github.com/insolar/insolar/network/hostnetwork/packet"
 	"github.com/insolar/insolar/network/pulsenetwork"
+	"github.com/insolar/insolar/pulse"
 
 	"github.com/stretchr/testify/require"
 )
