+++ network/consensus/tests/pulse_test.go
@@ -55,7 +55,7 @@ import (
 	"time"
 
 	"github.com/insolar/insolar/network/consensus/common/longbits"
-	"github.com/insolar/insolar/network/consensus/common/pulse"
+	"github.com/insolar/insolar/pulse"
 )
 
 func randBits256() longbits.Bits256 {
